\name{acousaR-package}
\alias{acousaR-package}
\alias{acousaR}
\docType{package}
\title{
Analysis of acoustic survey data in R
}
\description{
The International Council for the Exploration of the Sea runs various acoustic surveys that cover the Baltic Sea, Skagerrak, Kattegat, North Sea, English Channel, Celtic Sea, Irish Sea, Bay of Biscay and the eastern Atlantic from the Shetlands to Gibraltar. This library enables the data to be analyzed in R. 
}
\details{
\tabular{ll}{
Package: \tab acousaR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-03-19\cr
License: \tab GNU GPL v2\cr
}
Run the package using the ACOUsticSurveyAnalysis.r script
}
\author{
Sascha Fassler
Maintainer: Sascha Fassler <sascha.fassler@wur.nl>
}
\references{
Simmonds, E.J., MacLennan, D.N. (2005). Fisheries acoustics: theory and practice, 2nd edition. Blackwell Publishing, Oxford.
}
\keyword{ 
acoustic surveys
ICES }
\seealso{
\code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
